Split mock objects
